规格改变后主控区底图更新

+++ skyInnerSX/skyInnerSX/Popover Views/Settings Popover View/skySettingConfigVC.h
@@ -25,7 +25,7 @@
 @protocol skySettingConfigVCDelegate <NSObject>
 
 // 设定行列数
-- (void)setScreenRow:(NSInteger)nRow andColumn:(NSInteger)nColumn;
+- (void)setScreenRow:(int)nRow andColumn:(int)nColumn;
 
 @end
